pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"

[project]
name = "leo-benchmarks"
version = "1.0.0"
description = "Leo — A comprehensive AI model benchmarking platform. Evaluate any language model across 60+ benchmarks with a single command."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Pradyumn Tandon", email = "pradyumn@vrip7.com"},
]
maintainers = [
    {name = "VRIP7", email = "contact@vrip7.com"},
]
keywords = [
    "ai", "benchmarks", "llm", "evaluation", "machine-learning",
    "deep-learning", "transformers", "language-models", "mlops",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.38.0",
    "accelerate>=0.27.0",
    "datasets>=2.18.0",
    "tokenizers>=0.15.0",
    "safetensors>=0.4.0",
    "huggingface-hub>=0.21.0",
    "sentencepiece>=0.1.99",
    "protobuf>=4.25.0",
    "lm-eval>=0.4.0",
    "peft>=0.9.0",
    "bitsandbytes>=0.42.0",
    "scipy>=1.12.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "psutil>=5.9.0",
    "pynvml>=11.5.0",
    "filelock>=3.13.0",
]

[project.optional-dependencies]
unsloth = [
    "unsloth>=2024.1",
]
vllm = [
    "vllm>=0.3.0",
]
gguf = [
    "llama-cpp-python>=0.2.0",
]
all = [
    "leo-benchmarks[unsloth,vllm,gguf]",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.urls]
Homepage = "https://vrip7.com"
Repository = "https://github.com/vrip7/leo"
Documentation = "https://github.com/vrip7/leo/wiki"
"Bug Tracker" = "https://github.com/vrip7/leo/issues"
"Author" = "https://pradyumntandon.com"

[project.scripts]
leo = "leo.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["leo*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
